initially add batch and parallel

+++ molop/io/bases/file_base.py
@@ -2,16 +2,17 @@
 Author: TMJ
 Date: 2024-01-07 13:50:55
 LastEditors: TMJ
-LastEditTime: 2024-01-11 09:56:24
+LastEditTime: 2024-01-25 22:56:04
 Description: 请填写简介
 """
 import os
-from typing import List, Union, Tuple, TypeVar
+from typing import List, Tuple, TypeVar, Union
 
 from molop.io.bases.molblock_base import BaseBlockParser, QMBaseBlockParser
 from molop.io.coords_file.GJFBlockParser import GJFBlockParser
 from molop.io.coords_file.SDFBlockParser import SDFBlockParser
 from molop.io.coords_file.XYZBlockParser import XYZBlockParser
+from molop.io.qm_file.G16FCHKBlockParser import G16FCHKBlockParser
 from molop.io.qm_file.G16IRCBlockParser import G16IRCBlockParser
 from molop.io.qm_file.G16LOGBlockParser import G16LOGBlockParser
 from molop.io.qm_file.XTBOUTBlockParser import XTBOUTBlockParser
@@ -22,6 +23,7 @@ BlockType = Union[
     XYZBlockParser,
     G16LOGBlockParser,
     G16IRCBlockParser,
+    G16FCHKBlockParser,
     XTBOUTBlockParser,
 ]
 
@@ -52,6 +54,7 @@ class BaseFileParser:
                 raise ValueError(
                     f"File format {file_format} not in {self._allowed_formats}"
                 )
+        self._file_format = file_format
 
     def __iter__(self):
         self.__index = 0
@@ -86,6 +89,10 @@ class BaseFileParser:
         return self._file_path
 
     @property
+    def file_name(self) -> str:
+        return os.path.basename(self._file_path)
+
+    @property
     def frames(
         self,
     ) -> List[BlockType]:
@@ -108,14 +115,39 @@ class BaseFileParser:
     def to_SDF_block(self) -> str:
         return "$$$$\n".join([frame.to_SDF_block() for frame in self.__frames])
 
-    def to_XYZ_file(self, file_path: str):
+    def to_GJF_block(self, prefix: str = None, suffix="\n\n") -> str:
+        """Only extract the last frame."""
+        return self.__frames[-1].to_GJF_block(prefix=prefix, suffix=suffix)
+
+    def to_XYZ_file(self, file_path: str = None):
+        if file_path is None:
+            file_path = self._file_path
+        if not os.path.isfile(file_path):
+            raise IsADirectoryError(f"{file_path} is not a file.")
+        file_path = os.path.splitext(file_path)[0] + ".xyz"
         with open(file_path, "w") as f:
             f.write(self.to_XYZ_block())
         f.close()
 
-    def to_SDF_file(self, file_path: str):
+    def to_SDF_file(self, file_path: str = None):
+        if file_path is None:
+            file_path = self._file_path
+        if not os.path.isfile(file_path):
+            raise IsADirectoryError(f"{file_path} is not a file.")
+        file_path = os.path.splitext(file_path)[0] + ".sdf"
         with open(file_path, "w") as f:
             f.write(self.to_SDF_block())
+        f.close()
+
+    def to_GJF_file(self, file_path: str = None, prefix: str = None, suffix="\n\n"):
+        """Only extract the last frame."""
+        if file_path is None:
+            file_path = self._file_path
+        if not os.path.isfile(file_path):
+            raise IsADirectoryError(f"{file_path} is not a file.")
+        file_path = os.path.splitext(file_path)[0] + ".gjf"
+        with open(file_path, "w") as f:
+            f.write(self.to_GJF_block(prefix=prefix, suffix=suffix))
         f.close()
 
     def summary(self):
